Split GetSeqDepth.sh scripts into two
One obtains sequencing depth using only the bam file and the other one uses a fixed read length.
For the program will be probably easier to use the one in which you do not need to add the read length either because read length are different among samples or because read lengths are variable.